Changes to allow diffing of tables versions
Added a diff function to the Table object to allow output only the
difference between two tables to be output as a new table. Clones
the argument table and contains any data not present in the table
diff was called on. There is also a utility script to run the diff on
two tables.

Removed the converter scripts for the WMO and NCEP tables, and created
a combined script.

Removed the FlagTableDefinition is combine into the CodeFlagDefinition
object.

Added a function to output a dict from any BUFRTableObject that is
compatable with a Pandas DataFrame.